Changes Cache & Styles, working in Links
