+++ client/src/Styles/ProductCard_style.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+export const ProdCard = styled.div `
+  display: flex;
+
+
+
+
+
+
+
+
+
+`
