+++ client/src/Styles/styles.css
@@ -1,0 +1,176 @@
+/* .productCard{
+} */
+
+/* ESTILO CARTA */
+.Card {
+  display: flex;
+  padding: 10px;
+  border-radius: 0.5rem;
+  margin: 2rem;
+  flex-direction: column;
+  border: 0.1rem solid #DDD;
+  box-shadow: 0.3rem 0.3rem 0.5rem #AAA;
+  transition: transform 0.5s;
+}
+
+.card-img {
+  padding-top: 10px;
+  width: 100%;
+}
+
+.Card:hover {
+  transform: scale(1.1);
+}
+
+.card-body {
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+  justify-content: center;
+}
+
+.card-text {
+  display: flex;
+  justify-content: center;
+}
+
+
+/* ESTILO CARTAS (catalogo) */
+.cardSize {
+  display: flex;
+  flex-grow: 1;
+  
+}
+
+.catalogueDiv {
+  display: inline-flex;
+  flex-wrap: wrap;
+}
+
+
+/* BARRA BUSQUEDA */
+.searchBar{
+  padding: 10px;
+  display: flex;
+  justify-content: flex-end;
+  background-color:bisque;
+}
+
+.cssDeLaBarritaDeBusqueda {
+  margin-right: 10px;
+}
+
+/* .button-submit {
+  
+}
+
+.button-submit:hover {
+
+} */
+
+
+/* ESTILO BOTON */
+.buttonAdd {
+  border-radius: 4px;
+  background-color: #6495ED;
+  border: none;
+  color: #FFFFFF;
+  text-align: center;
+  font-size: 10px;
+  padding: 10px;
+  width: 100px;
+  transition: all 0.5s;
+  cursor: pointer;
+  margin: 5px;
+}
+
+.buttonAdd span {
+  cursor: pointer;
+  display: inline-block;
+  position: relative;
+  transition: 0.5s;
+}
+
+.buttonAdd span:after {
+  content: '\0bb';
+  position: absolute;
+  opacity: 0;
+  top: 0;
+  right: -20px;
+  transition: 0.5s;
+}
+
+.buttonAdd:hover span {
+  padding-right: 15px;
+}
+
+.buttonAdd:hover span:after {
+  opacity: 1;
+  right: 0;
+}  
+
+
+/* PAGINATION */
+
+.pagination{
+    display: flex; 
+    justify-content: flex-end;
+    margin-right: 3.5rem;
+}
+
+.page-list{
+    text-decoration: none;
+    list-style: none;
+    float: left;
+    color: black;
+    margin: 0.5rem;
+    padding: 8px 16px;
+    background-color: #ddd;
+    border-radius: .6rem;
+    border: none
+  }
+  
+.page-list a:active {
+    background-color: #4CAF50;
+    color: white;
+    border-radius: 5px;
+}
+
+.page-list:hover {
+    background-color: rgb(235, 199, 145);
+    border-radius: 5px;
+}
+.arrows{
+    color: rgb(62, 149, 179);
+    text-decoration: none;
+}
+
+
+/* FILTER */
+
+.filterDiv {
+  display: flex;
+  justify-content: flex-start;
+  /* justify-content: space-evenly; */
+  margin-top: 10px;
+}
+.order{
+    display: flex;
+    align-items: center;
+    margin-left: 1rem;
+}
+select{
+    background-color: rgb(241, 227, 227);
+    padding: 10px;
+    border-radius: 0.5rem;
+    margin: 1rem;
+    border: 0.05rem solid #DDD;
+    box-shadow: 0.3rem 0.3rem 0.5rem #AAA;
+    display: flex;
+}
+.input{
+    text-transform: capitalize;
+}
+.results{
+    margin:1rem;
+}
