+++ client/src/components/Cart.js
@@ -1,5 +1,3 @@
-// import React, { useState } from 'react'
-
 
 function Cart ({cartItems, removeFromCart}) {
