+++ client/src/components/Catalogue.js
@@ -1,9 +1,11 @@
 import ProductCard from './ProductCard';
+import styles from '../Styles/styles.css'
+
 
 function Catalogue ({ products , addToCart }) {
-  
+  console.log(styles)
   return (
-    <div>
+    <div className="catalogueDiv">
       {
         products.map((product) => {
           return (
