+++ client/src/components/Home.js
@@ -1,0 +1,22 @@
+// import React from "react";
+// import {
+//   BrowserRouter as Router,
+//   Switch,
+//   Route,
+//   Link
+// } from "react-router-dom";
+import Features from './Features';
+// import ShoppingPhoto from '../Styles/img/home.jpg'
+// Aca solo habra un link a la pagina de filtros y search y productos
+// y el diseño aca de la pagina principal
+
+
+function Home() {
+  return (
+    <div>
+      <Features />
+    </div>
+  )
+}
+
+export default Home;
